Flytta utledning av årak ikke tilgang tekst til backend.
Slik at den kan brukast i ExtendedApiError også.
